Quang vừa set lại các tên file nhé! file chính đựng tất cả các file của java sẽ là "heartBeat". bên trong đó có file "frame" là đựng các frame chúng ta code. Q có để luôn các file nền để ae pull về thì chạy luôn nó đỡ lỗi nha! sau này có các phần việc khác liên quan đến java thì sẽ tạo 1 folder mới trong folder "heartBeat" nha mọi người! chúc mọi người một ngày làm việc hiệu quả!
