+++ heartBeat/frame/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=69b5bc78
+build.xml.script.CRC32=2ad98cd9
+build.xml.stylesheet.CRC32=f85dc8f2@1.101.0.48
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=69b5bc78
+nbproject/build-impl.xml.script.CRC32=24bbf0ad
+nbproject/build-impl.xml.stylesheet.CRC32=12e0a6c2@1.101.0.48
