+++ heartBeat/frame/nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\TRAN VAN QUANG\\AppData\\Roaming\\NetBeans\\12.6\\build.properties
