+++ heartBeat/frame/nbproject/private/private.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
+    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
+    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
+        <group>
+            <file>file:/C:/Users/TRAN%20VAN%20QUANG/Documents/NetBeansProjects/FrontEnd/src/frame/dangKy.java</file>
+            <file>file:/C:/Users/TRAN%20VAN%20QUANG/Documents/NetBeansProjects/FrontEnd/src/quanLyUser/quanLyNguoiDung.java</file>
+            <file>file:/C:/Users/TRAN%20VAN%20QUANG/Documents/NetBeansProjects/FrontEnd/src/frame/dangNhap.java</file>
+            <file>file:/C:/Users/TRAN%20VAN%20QUANG/Documents/NetBeansProjects/FrontEnd/src/frame/frame.java</file>
+            <file>file:/C:/Users/TRAN%20VAN%20QUANG/Documents/NetBeansProjects/FrontEnd/src/quanLyUser/dialogXoa.java</file>
+        </group>
+    </open-files>
+</project-private>
